Article Entity 생성 및 Long 컬럼 변경
Contents
- 과도한 범위의 타입 Long을 int로 변경. Integer까지도 불필요해 보여 int로 수정
- 학습을 위해 기본 애노테이션을 모두 Bean으로 변경

Todo
- Article html 페이지 수정

+++ src/main/java/com/kakao/cafe/core/domain/member/Member.java
@@ -6,17 +6,17 @@ import java.util.Objects;
 
 public class Member {
 
-    private Long id;
+    private int id;
     private String email;
     private String password;
     private String nickName;
     private LocalDateTime createAt;
 
-    public Member (Long id){
+    public Member (int id){
         this.id = id;
     }
 
-    public Member(Long id, String email, String password, String nickName) {
+    public Member(int id, String email, String password, String nickName) {
         this.id = id;
         this.email = email;
         this.password = password;
@@ -33,7 +33,7 @@ public class Member {
         return LocalDateTime.now();
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
@@ -58,7 +58,7 @@ public class Member {
     }
 
     public static class Builder {
-        private Long id;
+        private int id;
         private String email;
         private String password;
         private String nickName;
@@ -72,7 +72,7 @@ public class Member {
             this.createAt = member.createAt;
         }
 
-        public Builder id(Long id){
+        public Builder id(int id){
             this.id = id;
             return this;
         }
