+++ src/main/java/com/kakao/cafe/core/repository/member/SequenceGenerator.java
@@ -1,11 +1,11 @@
 package com.kakao.cafe.core.repository.member;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class SequenceGenerator {
-    private static final AtomicLong sequence = new AtomicLong(10);
+    private static final AtomicInteger sequence = new AtomicInteger(10);
 
-    public static long getSequence(){
+    public static int getSequence(){
         return sequence.incrementAndGet();
     }
 }
